Add any and filter enumerables
