+++ README.md
@@ -1,2 +1,53 @@
 # enumerable-ruby
 Usage of Ruby module inside a class. Project provides a class MyList and a module MyEnumerable.
+
+## Clone repo to run the program and decode the message!
+
+## Built With
+- Ruby
+- irb
+- CMD with Ruby
+
+## Getting Started
+
+To get a copy of this project running on your local computer, do the followings:
+
+### Prerequisites
+- Must have a working computer
+- A Web-Browser installed (Chrome - Recommended)
+
+### Setup
+- Install a text Editor. For example [_VS_](https://visualstudio.microsoft.com/)
+- Clone this project repository on your local computer. [link](https://github.com/Kamba56/Morse-Decode)
+- Open the project directory with _Text Editor_
+
+### Install
+    [Ruby](https://rubyinstaller.org/) 
+
+
+## Authors
+
+👤 **Nadya**
+
+- GitHub: [@petushka1](https://github.com/petushka1)
+- LinkedIn: [LinkedIn](http://www.linkedin.com/in/nadyalevina)
+
+## Collaborators
+
+👤 **Toma John Musa**
+
+- GitHub: [@Kamba56](https://github.com/Kamba56)
+- Twitter: [@Kamba_TJ](https://twitter.com/Kamba_TJ)
+- LinkedIn: [Toma](https://linkedin.com/in/kamba56)
+
+
+## :handshake: Contributing
+
+Contributions, issues, and feature requests are welcome!
+Feel free to check the [issues page](../../issues/).
+
+## Show your support
+Give a :star:️ if you like this project!
+
+## Acknowledgments
+- Fellow Micronauts
