+++ mylist.rb
@@ -1,0 +1,24 @@
+require 'enumerable_spec'
+# Class with instance variable impelmented
+
+class MyList
+  include MyEnumerable
+
+  def initialize(*number)
+    @list = number
+  end
+
+  def each(&block)
+    yield(@list)
+    @list.all? &block
+  end
+end
+
+list = MyList.new(1, 2, 3, 4)
+list.all? { |e| e < 5 }
+list.all? { |e| e > 5 }
+
+list.any? { |e| e == 2 }
+list.any? { |e| e == 5 }
+
+list.filter { |e| e.even? }
